Добавил мягкое удаление Материалов и Поставщиков, а так же добавил проверку чтобы нельзя было удалить Материалы из Поставщиков если по ним уже были движения в системе. Добавил в систему вывод сообщения об ошибке.

+++ app/Http/Controllers/MaterialController.php
@@ -91,6 +91,11 @@ class MaterialController extends Controller
      */
     public function destroy(Material $material)
     {
+        if ($material->movementMaterials()->count() > 0) {
+            return redirect()->route('materials.index')
+                ->with('error', 'Невозможно удалить материал, так как он используется в системе');
+        }
+
         $material->delete();
 
         return redirect()->route('materials.index')->with('success', 'Материал удален');
